with some help from other students I managed to figure out how to get this darnthing working.
