+++ friends/src/components/friendform.js
@@ -33,33 +33,6 @@ class FriendForm extends Component {
             .catch((err) => console.log)
     }
     
-  // example edit function with put request
-  editFriend = (id) => {
-    const updatedFriendObj = {
-      name: this.state.name,
-      age: this.state.age,
-      email: this.state.email
-    }
-    axios.put(`http://localhost:5000/friends/${id}`, updatedFriendObj)
-    .then(response => {
-      this.setState({
-        friends: response.data
-      })
-    })
-    .catch((err) => console.log(err))
-  }
-
-  //  example delete function with delete request
-   deleteFriend = (id) => {
-     axios.delete(`http://localhost:5000/friends/${id}`)
-     .then(response => {
-       this.setState({
-         friends: response.data
-       })
-     })
-     .catch((err) => console.log(err))
-   }
-
     render() {
         return (  
         <div>
